refactor(wal): use 32-bit txn_id, simplify commit, add --use_wal flag
Change txn_id from u64 to u32 in WAL records and related structures.
Remove the txn_id parameter from _WalWriter::commit() (it is already known from begin()).
Adjust _WalWriter::begin() to store the last commit id before writing the record.
Update the benchmark tool to expose a --use_wal flag for testing with write-ahead logging enabled/disabled.

+++ include/leaves/intern/db/_cursor.hpp
@@ -605,7 +605,7 @@ struct _WalAware {
 
   void _wal_close() { _use_wal = false; }
 
-  void _wal_begin(uint64_t txn_id) {
+  void _wal_begin(uint32_t txn_id) {
     if (_use_wal) _derived()._db->wal_begin(txn_id);
   }
 
@@ -621,7 +621,7 @@ struct _WalAware {
     if (_use_wal) _derived()._db->wal_prepare();
   }
 
-  void _wal_commit(uint64_t txn_id) {
+  void _wal_commit(uint32_t txn_id) {
     if (_use_wal) _derived()._db->wal_commit(txn_id);
   }
